# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Casa de Apuestas API",
        "description": "API para casa de apuestas: autenticación, usuarios, deportes, eventos, cuotas, wallet y apuestas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API versión 1"
        }
    ],
    "paths": {
...
        "/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Listar eventos deportivos disponibles",
                "description": "Lista eventos activos y disponibles para usuarios autenticados.",
                "operationId": "08bb4f6a89011fba0aa159a9c841c2b5",
                "parameters": [
                    {
                        "name": "sport_key",
                        "in": "query",
                        "description": "Clave del deporte",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "soccer_australia_aleague"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Estado del evento",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "scheduled",
                                "live"
                            ]
                        },
                        "example": "scheduled"
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Fecha inicial",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-05-01"
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Fecha final",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-05-31"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 15
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eventos deportivos obtenidos correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/{sportEvent}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Ver detalle de un evento deportivo",
                "description": "Muestra el detalle de un evento, estado, etiqueta en vivo/cerrado y cuotas disponibles.",
                "operationId": "3d46ae63d76dd3b477a565f9ca2889d5",
                "parameters": [
                    {
                        "name": "sportEvent",
                        "in": "path",
                        "description": "ID interno del evento deportivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del evento obtenido correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/events/sync": {
            "post": {
                "tags": [
                    "Admin Events"
                ],
                "summary": "Sincronizar eventos deportivos",
                "description": "Consulta eventos desde The Odds API por deporte activo y los guarda o actualiza.",
                "operationId": "cfe67fee6eab6c375ba35604a86a52d0",
                "parameters": [
                    {
                        "name": "sport_key",
                        "in": "query",
                        "description": "Sincronizar eventos de un deporte específico",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "soccer_australia_aleague"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eventos deportivos sincronizados correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error al sincronizar eventos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/events/statuses/sync": {
            "post": {
                "tags": [
                    "Admin Events"
                ],
                "summary": "Actualizar estados de eventos deportivos",
                "description": "Actualiza automáticamente estados de eventos: scheduled, live, completed o cancelled según datos disponibles.",
                "operationId": "c0b7c96d6192d4b578b462197372ce20",
                "parameters": [
                    {
                        "name": "sport_key",
                        "in": "query",
                        "description": "Actualizar estados de un deporte específico",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "soccer_australia_aleague"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estados de eventos actualizados correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error al actualizar estados"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token personal de acceso. Formato: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}